+++ src/Tree/Entity/Repository/ClosureTreeRepository.php
@@ -278,8 +278,8 @@ class ClosureTreeRepository extends AbstractTreeRepository
 
         try {
             foreach ($nodesToReparent as $nodeToReparent) {
-                $id = $meta->getReflectionProperty($pk)->getValue($nodeToReparent);
-                $meta->getReflectionProperty($config['parent'])->setValue($nodeToReparent, $parent);
+                $id = $meta->getFieldValue($nodeToReparent, $pk);
+                $meta->setFieldValue($nodeToReparent, $config['parent'], $parent);
 
                 $dql = "UPDATE {$config['useObjectClass']} node";
                 $dql .= " SET node.{$config['parent']} = :parent";