+++ CHANGELOG.md
@@ -21,6 +21,7 @@ a release.
 
 ### Changed
 - Sluggable: Replaced abandoned `behat/transliterator` with `symfony/string` for default transliteration and urlization steps (#2985)
+- Use `getFieldValue` and `setFieldValue` methods to support `doctrine/orm` >= 3.4 (#2966)
 
 ## [3.20.1] - 2025-09-14
 ### Fixed